Moved BLE functionality to its own module
